CI: Move miscellaneous tests into separate file
